requirements.txt
streamlit==1.38.0
pandas==2.1.1
plotly==6.1.0
requests==2.31.0
python-dotenv==1.0.1
streamlit-autorefresh==0.0.2
pytz==2025.7
sqlite3-binary==0.0.1
